Simulate genetic data from simple P. falciparum model

Simulate genetic data from a simple model of P. falciparum epidemiology and genetics.

sim_falciparum(
  a = 0.3,
  p = 0.9,
  mu = -log(p),
  u = 12,
  v = 10,
  g = 10,
  prob_infection = seq(0.1, 0.01, -0.01),
  duration_infection = dgeom(1:500, 1/200),
  infectivity = 1,
  max_innoculations = 5,
  H = 1000,
  seed_infections = 100,
  M = 1000,
  mig_matrix = diag(length(M)),
  L = 24,
  prob_cotransmission = 0.5,
  time_out = 100,
  report_progress = TRUE
)

Arguments

a
human blood feeding rate. The proportion of mosquitoes that feed on humans each day.

p
mosquito probability of surviving one day.

mu
mosquito instantaneous death rate. mu = -log(p) unless otherwise specified.

u
intrinsic incubation period. The number of days from infection to blood-stage infection in a human host.

v
extrinsic incubation period. The number of days from infection to becoming infectious in a mosquito.

g
lag time between human blood-stage infection and production of gametocytes.

prob_infection
probability a human becomes infected after being bitten by an infected mosquito.

duration_infection
vector specifying probability distribution of time (in days) of a malaria episode.

infectivity
probability a mosquito becomes infected after biting an infective human host.

max_innoculations
maximum number of innoculations that an individual can hold simultaneously.

H
human population size, which is assumed the same in every deme.

seed_infections
vector specifying the initial number of infected humans in each deme.

M
vector specifying mosquito population size (strictly the number of adult female mosquitoes) in each deme.

mig_matrix
migration matrix specifing the daily probability of migrating from each deme to each other deme. Migration must be equal in both directions, meaning this matrix must be symmetric.

L
number of loci. The maximum number of loci is 1000, as at higher numbers haplotypes begin to exceed integer representation (2^L).

prob_cotransmission
probability of mosquito transmitting multiple haplotypes to host.

time_out
vector of times (days) at which output is produced.

report_progress
if TRUE then a progress bar is printed to the console during simuation.