Square a vector of values

Simple test function that demonstrates some of the features of this package by squaring an input vector of values.

square(x = 1:5)

Arguments

x
vector of values.

Examples

# Find square of first 100 values
square(1:100)

#> running R square function

#> running C++ dummy1_cpp function

#> processing output

#>   [1]     1     4     9    16    25    36    49    64    81   100   121   144
#>  [13]   169   196   225   256   289   324   361   400   441   484   529   576
#>  [25]   625   676   729   784   841   900   961  1024  1089  1156  1225  1296
#>  [37]  1369  1444  1521  1600  1681  1764  1849  1936  2025  2116  2209  2304
#>  [49]  2401  2500  2601  2704  2809  2916  3025  3136  3249  3364  3481  3600
#>  [61]  3721  3844  3969  4096  4225  4356  4489  4624  4761  4900  5041  5184
#>  [73]  5329  5476  5625  5776  5929  6084  6241  6400  6561  6724  6889  7056
#>  [85]  7225  7396  7569  7744  7921  8100  8281  8464  8649  8836  9025  9216
#>  [97]  9409  9604  9801 10000